requirements.txt
# Z-Dungeon Core 2.0 - Requirements
requests>=2.28.0
pyyaml>=6.0
rich>=12.0.0
prompt_toolkit>=3.0.29
jsonschema>=4.17.0
ollama>=0.1.0
python-dotenv>=0.21.0
fastapi>=0.95.0
uvicorn>=0.21.0
# Pour les tests et le développement
pytest>=7.3.1
black>=23.3.0